+++ sdk/resourcemanager/computeschedule/armcomputeschedule/models_serde.go
@@ -363,8 +363,8 @@ func (e *ExecuteCreateContent) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-// MarshalJSON implements the json.Marshaller interface for type ExecuteCreateFlexRequest.
-func (e ExecuteCreateFlexRequest) MarshalJSON() ([]byte, error) {
+// MarshalJSON implements the json.Marshaller interface for type ExecuteCreateFlexContent.
+func (e ExecuteCreateFlexContent) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "correlationid", e.Correlationid)
 	populate(objectMap, "executionParameters", e.ExecutionParameters)
@@ -372,8 +372,8 @@ func (e ExecuteCreateFlexRequest) MarshalJSON() ([]byte, error) {
 	return json.Marshal(objectMap)
 }
 
-// UnmarshalJSON implements the json.Unmarshaller interface for type ExecuteCreateFlexRequest.
-func (e *ExecuteCreateFlexRequest) UnmarshalJSON(data []byte) error {
+// UnmarshalJSON implements the json.Unmarshaller interface for type ExecuteCreateFlexContent.
+func (e *ExecuteCreateFlexContent) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
 		return fmt.Errorf("unmarshalling type %T: %v", e, err)